feat: Implement Audio Metadata Extraction and Editing Component
- Added `AudioMetaExtractor` component for extracting and editing audio metadata.
- Integrated file upload functionality with automatic submission on file selection.
- Implemented API calls for metadata extraction with error handling and user notifications.
- Introduced `AudioMetadataEditor` for editing extracted metadata.
- Enhanced `SearchLyrics` component with modal integration for improved user experience.
- Refactored `AudioFileConverter` page to utilize a dedicated component for audio conversion.
- Updated validation schema to improve error messaging for lyrics search.
- Improved UI elements for better accessibility and responsiveness.
